+++ src/components/detail/CommentList.tsx
@@ -1,11 +1,6 @@
-import React, { Fragment, useState } from 'react'
 import styled from '@emotion/styled'
-import { Modal, Avatar } from '@components/common'
 import { Comment, User } from '@interfaces'
-import { BiDotsHorizontalRounded, BiTrash } from 'react-icons/bi'
-import { getDate } from '@utils/getDate'
-import { useDeleteCommentMutation } from '@hooks/mutations/useDeleteCommentMutation'
-import { useRouter } from 'next/router'
+import CommentBox from '@components/detail/CommentBox'
 
 interface Props {
   user: User
@@ -13,93 +8,22 @@ interface Props {
 }
 
 const CommentList = ({ user, commentList }: Props) => {
-  const [isModalOpen, setIsModalOpen] = useState(false)
-  const [commentId, setCommentId] = useState(0)
-  const { mutate: deleteComment } = useDeleteCommentMutation()
-  const router = useRouter()
-
-  const handleUserClick = (userId: number | null) => {
-    if (!userId) {
-      return
-    }
-    router.push(`/user/${userId}`)
-  }
-
-  const handleCommentModalClick = (id: number) => {
-    setIsModalOpen(true)
-    setCommentId(id)
-  }
-
-  const handleCommentModalClose = () => {
-    setIsModalOpen(false)
-  }
-
-  const handleDeleteCommentClick = () => {
-    deleteComment(
-      {
-        commentId
-      },
-      {
-        onSuccess: () => {
-          setIsModalOpen(false)
-        }
-      }
-    )
-  }
-
   return (
-    <>
-      <CommentListContainer>
-        <CommnetCountText>댓글 {commentList.length}개</CommnetCountText>
-        {commentList.map(({ id, author, createdAt, comment }) => (
-          <Fragment key={id}>
-            <CommentWrapper>
-              <Avatar
-                size={4}
-                src={author.image}
-                isLoading={false}
-                onClick={() => handleUserClick(author.id)}
-              />
-              <CommentContainer>
-                <CommentHeader>
-                  <NameAndDateWrapper>
-                    <UserNameText onClick={() => handleUserClick(author.id)}>
-                      {author.nickName}
-                    </UserNameText>
-                    <DateText>{getDate(createdAt)}</DateText>
-                  </NameAndDateWrapper>
-                  {user.id === author.id && (
-                    <>
-                      <Dots
-                        size={20}
-                        onClick={() => handleCommentModalClick(id)}
-                      />
-                    </>
-                  )}
-                </CommentHeader>
-                <CommentText>{comment}</CommentText>
-              </CommentContainer>
-            </CommentWrapper>
-          </Fragment>
-        ))}
-      </CommentListContainer>
-      <Modal
-        visible={isModalOpen}
-        onClose={handleCommentModalClose}
-        option="drawer"
-      >
-        <ModalItem onClick={handleDeleteCommentClick}>
-          <BiTrash size={25} />
-          삭제
-        </ModalItem>
-      </Modal>
-    </>
+    <CommentListContainer>
+      <CommnetCountText>댓글 {commentList.length}개</CommnetCountText>
+      {commentList.map(({ id, author, createdAt, comment }) => (
+        <CommentBox
+          key={id}
+          id={id}
+          author={author}
+          createdAt={createdAt}
+          comment={comment}
+          user={user}
+        />
+      ))}
+    </CommentListContainer>
   )
 }
-
-const Flex = styled.div`
-  display: flex;
-`
 
 const CommentListContainer = styled.div``
 
@@ -107,61 +31,6 @@ const CommnetCountText = styled.div`
   font-size: 1.6rem;
   font-weight: 700;
   margin-bottom: 2rem;
-`
-
-const CommentWrapper = styled(Flex)`
-  margin-bottom: 1rem;
-  padding-bottom: 1rem;
-  border-bottom: 0.1rem solid #f1f1f1;
-`
-
-const CommentContainer = styled.div`
-  width: calc(100% - 4rem);
-  margin-left: 1rem;
-  padding-bottom: 0.8rem;
-`
-
-const CommentHeader = styled(Flex)`
-  justify-content: space-between;
-  margin-bottom: 0.4rem;
-`
-
-const NameAndDateWrapper = styled(Flex)`
-  align-items: center;
-`
-
-const UserNameText = styled.div`
-  font-size: 1.6rem;
-  font-weight: 700;
-  cursor: pointer;
-`
-
-const DateText = styled.span`
-  margin-left: 1.4rem;
-  color: #a3a3a3;
-`
-
-const Dots = styled(BiDotsHorizontalRounded)`
-  justify-self: end;
-  cursor: pointer;
-`
-
-const CommentText = styled.div`
-  font-size: 1.6rem;
-`
-
-const ModalItem = styled(Flex)`
-  align-items: center;
-  height: 5rem;
-  font-size: 2rem;
-  justify-content: center;
-  width: 100vw;
-  max-width: 50rem;
-  cursor: pointer;
-
-  &:not(:first-of-type) {
-    margin-top: 1rem;
-  }
 `
 
 export default CommentList
